agent/requirements.txt
strands-agents
strands-agents-tools
strands-agents-builder
strands-agents[otel]
langfuse
python-dotenv
uv